pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "copilot-sdk-proxy"
version = "0.1.0"
description = "OpenAI-compatible proxy for GitHub Copilot SDK"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "github-copilot-sdk @ git+https://github.com/github/copilot-sdk.git#subdirectory=python",
    "python-dotenv>=1.0",
]

[project.scripts]
copilot-proxy = "app.main:run"
